Test assignValue and populateGameBoard functions
